requirements.txt
pika==1.1.0
rethink==2.4.7
flask==1.1.2
flask-restful==0.3.8
flask-marshmallow==0.12.0
flask-bcrypt==0.7.1